add policies for lessons and tokens, create pages for 404 and 403 errors

+++ app/Http/Controllers/TokenController.php
@@ -2,15 +2,32 @@
 
 namespace Étula\Http\Controllers;
 
-use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Étula\LessonToken;
 use Étula\Lesson;
-use Étula\LessonTeacher;
 
 class TokenController extends Controller
 {
+    /**
+     * Create new Controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        // TODO: Determine if admins and students need to do something with tokens, if they do we need policies.
+        // middleware = for all class, policy = can differ for some actions
+        $this->middleware(['auth', 'teacher']);
+    }
+
+    /**
+     * @param $id
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function create($id){
+        $this->authorize('create', $id);
+
         $token = new LessonToken;
         $token->lesson_id = $id;
 
@@ -31,12 +48,26 @@ class TokenController extends Controller
         return view('code', compact('token'));
     }
 
+    /**
+     * @param $id
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function delete($id){
-        LessonToken::where('lesson_id',$id)->delete();
+        $lt = LessonToken::where('lesson_id', $id)->first();
+        $this->authorize('delete', $lt);
+        $lt->delete();
         return redirect()->route('home');
     }
 
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function accept(Request $request){
+        $this->authorize('accept');
+
         $request->validate([
             'token' => 'required|numeric|digits:6'
         ]);
@@ -62,7 +93,14 @@ class TokenController extends Controller
         return redirect()->route('home');
     }
 
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function acceptByTeacher(Request $request){
+        $this->authorize('acceptByTeacher');
+
         $lesson = Lesson::where('id', $request->input('lesson_id'))->first();
         $lesson->presentStudents()->detach();
         foreach($_POST as $key=>$valeur){
